- Avoid `-Wconversion` warnings.


git-svn-id: http://svn.boost.org/svn/boost/branches/release@67664 b8fc166d-592f-0410-95f2-cb63ce0dd405

+++ libs/functional/hash/test/hash_complex_test.cpp
@@ -69,7 +69,7 @@ void complex_float_tests(Float*)
     generic_complex_tests(complex(0.5,0));
     generic_complex_tests(complex(25,0));
     generic_complex_tests(complex(25,0));
-    generic_complex_tests(complex(-67.5324535,56.23578678));
+    generic_complex_tests(complex(static_cast<Float>(-67.5324535),static_cast<Float>(56.23578678)));
 }
 
 template <class Integer>
